option to remove typefully buttons + important fixes
