+++ content-scripts/src/modules/options/navigation.js
@@ -1,8 +1,6 @@
-import { showGrowTab } from "../grow-tab";
 import svgAssets from "../svgAssets";
 import addStyles from "../utilities/addStyles";
 import removeElement from "../utilities/removeElement";
-import { getStorage } from "../utilities/storage";
 
 // Function to change Home Button
 export const changeHomeButton = (homeButton) => {
@@ -206,51 +204,6 @@ export const addListsButton = () => {
       listsButton.firstChild.lastChild.firstChild.innerText = "Lists";
       profileNode.insertAdjacentElement("beforebegin", listsButton);
     }
-  }
-};
-
-// Function to change Grow Button
-export const changeGrowButton = (growButton) => {
-  switch (growButton) {
-    case "off":
-      removeElement("mt-typefullyGrowButton");
-    case "on":
-      addGrowButton();
-      break;
-  }
-};
-
-// Function to add Grow button
-export const addGrowButton = async () => {
-  const userSetting = await getStorage("typefullyGrowTab");
-  if (userSetting !== "on") return;
-  const exists = document.querySelector("#mt-typefullyGrowButton");
-  if (exists) return;
-
-  const profileNode = document.querySelector(
-    'a[role="link"][data-testid="AppTabBar_Profile_Link"]'
-  );
-
-  if (profileNode) {
-    const growButton = document.createElement("div");
-    growButton.innerHTML = profileNode.innerHTML;
-
-    growButton.id = "mt-typefullyGrowButton";
-    growButton.ariaLabel = "Typefully Grow";
-    growButton.style.cursor = "pointer";
-    growButton.firstChild.firstChild.firstChild.innerHTML =
-      svgAssets.grow.normal;
-
-    growButton.onclick = () => {
-      growButton.firstChild.firstChild.firstChild.innerHTML =
-        svgAssets.grow.selected;
-      if (growButton.querySelector("span"))
-        growButton.querySelector("span").style.fontWeight = "700";
-      showGrowTab();
-    };
-
-    growButton.firstChild.lastChild.firstChild.innerText = "Grow";
-    profileNode.insertAdjacentElement("beforebegin", growButton);
   }
 };
