+++ content-scripts/src/modules/options/writer-mode.js
@@ -16,49 +16,53 @@ export const changeWriterMode = (writerMode) => {
   ) {
     switch (writerMode) {
       case "on":
-        addStyles(
-          "mt-writerMode",
-          `
-          body {
-            padding-left: 0 !important;
-          }
-          main[role="main"] > div {
-            width: 100% !important;
-            max-width: 100% !important;
-          }
-          header[role="banner"], 
-          [data-testid="sidebarColumn"],
-          [data-testid="primaryColumn"] > div > div:not(:nth-of-type(1)):not(:nth-of-type(2)):not(:nth-of-type(3)) {
-            display: none !important;
-          }
-          [data-testid="primaryColumn"] > div > div:nth-of-type(1) {
-            visibility: hidden !important;
-          }
-          div[data-testid="primaryColumn"] {
-            border-style: hidden !important;
-            padding-top: 3vh !important;
-            margin: 0 auto;
-          }
-          div[aria-labelledby="modal-header"][role="dialog"] {
-            width: 100vw !important;
-            max-width: 100vw !important;
-            top: 0 !important;
-            border-radius: 0 !important;
-          }
-          div[role="group"] > div:empty {
-            background-color: var(--body-bg-color) !important;
-          }
-          div[aria-labelledby="modal-header"][role="dialog"] > div {
-            border-radius: 0 !important;
-          }
-          div[aria-labelledby="modal-header"][role="dialog"] > div > div > div {
-            padding-bottom: 10vh !important;
-          }
-          `
-        );
+        setTimeout(() => {
+          addStyles(
+            "mt-writerMode",
+            `
+            body {
+              padding-left: 0 !important;
+            }
+            main[role="main"] > div {
+              width: 100% !important;
+              max-width: 100% !important;
+            }
+            header[role="banner"], 
+            [data-testid="sidebarColumn"],
+            [data-testid="primaryColumn"] > div > div:not(:nth-of-type(1)):not(:nth-of-type(2)):not(:nth-of-type(3)) {
+              display: none !important;
+            }
+            [data-testid="primaryColumn"] > div > div:nth-of-type(1) {
+              visibility: hidden !important;
+            }
+            div[data-testid="primaryColumn"] {
+              border-style: hidden !important;
+              padding-top: 3vh !important;
+              margin: 0 auto;
+            }
+            div[aria-labelledby="modal-header"][role="dialog"] {
+              width: 100vw !important;
+              max-width: 100vw !important;
+              top: 0 !important;
+              border-radius: 0 !important;
+            }
+            div[role="group"] > div:empty {
+              background-color: var(--body-bg-color) !important;
+            }
+            div[aria-labelledby="modal-header"][role="dialog"] > div {
+              border-radius: 0 !important;
+            }
+            div[aria-labelledby="modal-header"][role="dialog"] > div > div > div {
+              padding-bottom: 10vh !important;
+            }
+            `
+          );
+        }, 500);
+
         setTimeout(() => {
           addTypefullyPlugToWriterMode();
         }, 1000);
+
         break;
 
       case "off":
@@ -122,6 +126,7 @@ export const removeTypefullyPlugFromWriterMode = () => {
     "typefully-writermode-link"
   );
   typefullyLinkElement && typefullyLinkElement.remove();
+
   const typefullyBox = document.getElementById("typefully-writermode-box");
   typefullyBox && typefullyBox.remove();
 };
